only choose from available action

+++ gailtf/baselines/ppo1/mlp_policy.py
@@ -133,7 +133,7 @@ class MlpPolicy(object):
         for i in range(0, self.available_action_size):
             if available_action[i] == 1.0:
                 available_act.append(i)
-        available_act = np.array(available_act)
+        available_act = np.array(available_act, dtype=np.int32)
 
         stochastic = U.get_placeholder(name="stochastic", dtype=tf.bool, shape=())
         ac = U.switch(stochastic, self.pd.sample(available_act), self.pd.mode(available_act))
